Rename serve command to dev command
