+++ packages/marko-web-cli/src/commands/index.js
@@ -1,11 +1,11 @@
 const build = require('./build');
 const newProject = require('./new');
-const serve = require('./serve');
+const dev = require('./dev');
 const lint = require('./lint');
 
 module.exports = (program) => {
   build(program);
-  serve(program);
+  dev(program);
   newProject(program);
   lint(program);
 };
